refactor(mitm-addon): split usage.py into package with per-connector dispatch
The 901-line usage.py accreted five unrelated concerns (counters, token
extraction, webhook delivery, model-provider reporting, X connector
billing).  Split into a usage/ package along those seams:

  usage/
    counters.py          flow + report counters, pending-path state
    extract.py           SSE / JSON Anthropic token parsers
    webhook.py           POST + retry + thread pool
    providers/
      model_provider.py  report_model_provider_usage
      connectors/
        __init__.py      report_connector_usage dispatcher
        x.py             X-specific NDJSON + billing (report_usage)

Per-connector dispatch table in providers/connectors/__init__.py applies
the universal run_id / firewall_billable / handler-registered gates once,
then delegates to x.report_usage (formerly self-gating).  Adding a new
billable connector becomes a new module plus a registration entry.

Naming tightened for consistency after the split:
- log_connector_usage -> report_connector_usage (it POSTs a billing
  webhook, not a log entry)
- x.log_usage -> x.report_usage (same reason; keeps the verb aligned
  across all three entry points)
- maybe_report_proxy_usage -> report_model_provider_usage ("proxy" was
  redundant, "model_provider" matches the firewall_name prefix and the
  module name; maybe_ prefix dropped now the docstring carries the
  no-op conditions)
- flow.metadata["proxy_usage"] -> flow.metadata["model_provider_usage"]
  for symmetry with flow.metadata["x_ndjson_state"]
- is_x_stream_path / create_x_ndjson_extractor -> is_stream_path /
  create_ndjson_extractor inside x.py; exposed as usage.x.is_stream_path
  etc. so caller sites keep the X context without stuttering
- Internal X symbols shed redundant _x_ / _X_ prefixes

build.rs walks mitm-addon/src recursively and embeds relative paths as
keys ("usage/counters.py", etc.).  proxy.rs creates parent directories
before writing so the runtime extractor rebuilds the package layout
before mitmdump loads the addon.

Tests retarget patch sites to the submodule that reads each name
(usage.webhook._opener, usage.providers.connectors.x.get_api_url, ...)
because Python's "from X import Y" creates module-local bindings that a
facade-level patch can't reach.  No test logic or assertions changed.
