+++ crates/runner/mitm-addon/src/mitm_addon.py
@@ -302,7 +302,7 @@ def responseheaders(flow: http.HTTPFlow) -> None:
     firewall_name = flow.metadata.get("firewall_name", "")
     is_model_provider = firewall_name.startswith("model-provider:")
     # Platform-billable firewall flag, sourced from vm_info["billableFirewalls"]
-    # via auth.handle_firewall_request.  Gates log_connector_usage (in response())
+    # via auth.handle_firewall_request.  Gates report_connector_usage (in response())
     # and the full-body response buffering that billing payload extraction needs.
     is_billable_flow = flow.metadata.get("firewall_billable", False)
     # X-specific NDJSON stream classification — tied to the x firewall itself,
@@ -313,12 +313,12 @@ def responseheaders(flow: http.HTTPFlow) -> None:
     # Classify X NDJSON streams early so we can register an incremental parser
     # and avoid buffering the (potentially multi-GB) stream body.  Only a
     # cheap path-only check happens here — full request metadata is parsed
-    # later in response() by log_connector_usage.
+    # later in response() by report_connector_usage.
     #
     # Gated on 2xx status so error responses (4xx/5xx on stream endpoints
     # return JSON, not NDJSON) fall through to the existing unbounded-buffer
     # path and preserve the full error body for forensic inspection in
-    # network logs.  log_connector_usage already skips non-2xx responses
+    # network logs.  report_connector_usage already skips non-2xx responses
     # so no billing record is affected either way.
     #
     # Reads ``original_url`` with no fallback — kept consistent with
@@ -329,7 +329,7 @@ def responseheaders(flow: http.HTTPFlow) -> None:
     is_x_stream = False
     if is_x_flow and _HTTP_STATUS_OK_MIN <= flow.response.status_code < _HTTP_STATUS_REDIRECT_MIN:
         stream_path = urllib.parse.urlparse(flow.metadata.get("original_url", "")).path
-        is_x_stream = usage.is_x_stream_path(stream_path)
+        is_x_stream = usage.x.is_stream_path(stream_path)
 
     # Track flows that will generate webhook POSTs (usage or connector billing)
     # so the runner can wait for them to reach response()/error() before stopping.
@@ -342,14 +342,14 @@ def responseheaders(flow: http.HTTPFlow) -> None:
         if "text/event-stream" in content_type:
             parser_fn, usage_dict = usage.create_sse_usage_extractor()
             sse_parser = parser_fn
-            flow.metadata["proxy_usage"] = usage_dict
+            flow.metadata["model_provider_usage"] = usage_dict
             sse_decompressor = body_utils.create_stream_decompressor(flow.response.headers)
     elif is_x_stream:
-        parser_fn, ndjson_state = usage.create_x_ndjson_extractor()
+        parser_fn, ndjson_state = usage.x.create_ndjson_extractor()
         ndjson_parser = parser_fn
-        # Deliberately NOT "proxy_usage" — that key would route through
-        # maybe_report_proxy_usage and trigger the model-provider webhook.
-        # x_ndjson_state is only consumed by log_connector_usage.
+        # Deliberately NOT "model_provider_usage" — that key would route through
+        # report_model_provider_usage and trigger the model-provider webhook.
+        # x_ndjson_state is only consumed by report_connector_usage.
         flow.metadata["x_ndjson_state"] = ndjson_state
         ndjson_decompressor = body_utils.create_stream_decompressor(flow.response.headers)
 
@@ -486,7 +486,7 @@ def response(flow: http.HTTPFlow) -> None:
     # Report proxy-extracted usage for model provider responses.
     # For non-streaming responses, fall back to extracting usage from the
     # buffered JSON body (buffer is never truncated for model providers).
-    if not flow.metadata.get("proxy_usage") and stream_buf:
+    if not flow.metadata.get("model_provider_usage") and stream_buf:
         firewall_name = flow.metadata.get("firewall_name", "")
         if firewall_name.startswith("model-provider:"):
             json_usage = usage.extract_usage_from_json(
@@ -494,11 +494,11 @@ def response(flow: http.HTTPFlow) -> None:
                 flow.response.headers if flow.response else None,
             )
             if json_usage:
-                flow.metadata["proxy_usage"] = json_usage
-    usage.maybe_report_proxy_usage(flow, run_id)
+                flow.metadata["model_provider_usage"] = json_usage
+    usage.report_model_provider_usage(flow, run_id)
 
     # Billable connector usage observation (issue #9504, stage 0).
-    usage.log_connector_usage(flow, run_id)
+    usage.report_connector_usage(flow, run_id)
 
     # Invalidate firewall header cache on 401 so next request gets fresh headers.
     # Also request a force-refresh so the next /firewall/auth fetch refreshes
@@ -582,17 +582,17 @@ def error(flow: http.HTTPFlow) -> None:
     log_network_entry(network_log_path, log_entry)
 
     # Report proxy-extracted usage for model provider responses.
-    # The SSE parser may have partially populated proxy_usage before the
+    # The SSE parser may have partially populated model_provider_usage before the
     # connection error occurred.  Partial data is better than none.
-    usage.maybe_report_proxy_usage(flow, run_id)
+    usage.report_model_provider_usage(flow, run_id)
 
     # Billable connector usage for X NDJSON streams that crash mid-flight
     # (issue #9534): the incremental parser populated x_ndjson_state during
     # chunks; log what was observed so partial streams aren't silently
-    # dropped from billing.  log_connector_usage no-ops when there's no
+    # dropped from billing.  report_connector_usage no-ops when there's no
     # response or the status is non-2xx, so normal model-provider errors
     # are unaffected.
-    usage.log_connector_usage(flow, run_id)
+    usage.report_connector_usage(flow, run_id)
 
     log_proxy_entry(
         proxy_log_path,
@@ -615,7 +615,7 @@ def done():
     ``shutdown(wait=True)`` blocks until all submitted futures complete;
     SIGKILL is the hard stop if any report takes too long.
     """
-    usage.usage_executor.shutdown(wait=True)
+    usage.webhook.usage_executor.shutdown(wait=True)
 
 
 # ============================================================================
